requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.1
GroupyAPI==0.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
Pillow==2.5.3
psutil==5.2.2
pyparsing==2.2.0
requests==2.3.0
six==1.10.0
Werkzeug==0.12.1
